minor edits: rename files and variables
