+++ micro_dl/input/__init__.py
@@ -1,8 +1,8 @@
 """Module for Input functions"""
 
 from .split_lif_stack import LifStackSplitter2D, LifStackSplitter3D
-from .correct_flat_field import FlatFieldEstimator2D
+from .estimate_flat_field import FlatFieldEstimator2D
 from .gen_crop_masks import MaskProcessor
-from .crop_stack import ImageStackTiler
+from .tile_stack import ImageStackTiler
 from .dataset import BaseDataSet, DataSetWithMask
 from .training_table import BaseTrainingTable, TrainingTableWithMask
